some work on debug commands
I added the Shell library but it seems it won't suit here since we'd
rather groovy to just die and syscall exec. But it seems that's not
something the jvm is good for.
I suppose that it'll be best to just move everything that doesn't need
the context out of the groovy script into bash.